all scenes are singleton
