--- src/main/java/org/example/hakmana/model/Mouse.java
@@ -1,12 +1,0 @@
-package org.example.hakmana.model;
-
-public class Mouse{
-    private String regNum;
-    private String model="No";
-    private String name;
-    private String status;
-    private String purchasedFrom;
-    private String Connectivity;
-
-
-}
