+++ src/main/java/org/example/hakmana/model/mainDevices/Laptops.java
@@ -1,10 +1,11 @@
-package org.example.hakmana.model;
+package org.example.hakmana.model.mainDevices;
 
 import javafx.scene.control.Alert;
 import javafx.scene.control.ButtonType;
 
 
 import lombok.*;
+import org.example.hakmana.model.DatabaseConnection;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
