+++ src/main/java/org/example/hakmana/model/mainDevices/PhotocpyMchine.java
@@ -1,8 +1,9 @@
-package org.example.hakmana.model;
+package org.example.hakmana.model.mainDevices;
 
 import javafx.scene.control.Alert;
 import javafx.scene.control.ButtonType;
-import lombok.*;
+
+import org.example.hakmana.model.DatabaseConnection;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -12,73 +13,59 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-public class Monitors extends Devices{
+
+public class PhotocpyMchine extends Devices {
     private DatabaseConnection conn=DatabaseConnection.getInstance();;
     private String regNum;
     private String model;
     private String status;
     private String userName;
-    @Setter
-    @Getter
-    private String screenSize;
-    private String purchasedFrom;
 
-    public Monitors(String regNum, String model, String userName, String status, String screenSize) {
+    public PhotocpyMchine(String regNum, String model, String userName, String status) {
         super(regNum, model, userName, status);
-        this.screenSize = screenSize;
     }
 
-    public Monitors(String regNum, String model, String userName, String status) {
-        super(regNum, model, userName,status);
-    }
-
-    public Monitors() {
-    }
-
-    @Override
-    public void setRegNum(String para1) {
-
+    public PhotocpyMchine() {
     }
 
     @Override
     public String getRegNum() {
-        return null;
+        return regNum;
     }
-
     @Override
-    public void setModel(String para1) {
-
+    public void setRegNum(String regNum) {
+        this.regNum = regNum;
     }
-
     @Override
     public String getModel() {
-        return null;
+        return model;
     }
-
     @Override
     public String getUserName() {
         return null;
     }
-
     @Override
-    public void setUserName(String para1) {
-
+    public void setModel(String model) {
+        this.model = model;
     }
-
-    @Override
-    public void setStatus(String para1) {
-
-    }
-
     @Override
     public String getStatus() {
-        return null;
+        return status;
+    }
+    @Override
+    public void setStatus(String status) {
+        this.status = status;
+    }
+    @Override
+    public void setUserName(String userName) {
+        this.userName = userName;
     }
 
-    public Monitors[] getDevices() {
-        List<Monitors> monitors = new ArrayList<>();
+    @Override
+    public PhotocpyMchine[] getDevices() {
+        List<PhotocpyMchine> photocopyMachines = new ArrayList<>();
         //pass query to the connection class
-        String sql = "SELECT Monitor.MonitorRegNum,Monitor.model,Monitor.status FROM Monitor";
+        String sql = "SELECT PhotoCopyMachine.* FROM PhotoCopyMachine";
 
         try {
             // get result set from connection class
@@ -86,26 +73,26 @@ public class Monitors extends Devices{
 
             // Iterate through the result set and create Desktop and DeviceUser objects
             while (resultSet.next()) {
-                Monitors monitor = new Monitors();
+                PhotocpyMchine photocopyMachine = new PhotocpyMchine();
 
-                monitor.setRegNum(resultSet.getString("MonitorRegNum"));
-                monitor.setModel(resultSet.getString("model"));
-                monitor.setStatus(resultSet.getString("status"));
-                monitor.setUserName("no user");
+                photocopyMachine.setRegNum(resultSet.getString("PhotoCopyMachineRegNum"));
+                photocopyMachine.setModel(resultSet.getString("model"));
+                photocopyMachine.setStatus(resultSet.getString("status"));
+                photocopyMachine.setUserName("no user");
 
-                monitors.add(monitor);//add monitor to the monitors list
+                photocopyMachines.add(photocopyMachine);//add photocopyMachines to the photocopyMachins list
             }
         }
         catch (SQLException e){
             throw new RuntimeException(e);
         }
 
-        return monitors.toArray(new Monitors[0]);
+        return photocopyMachines.toArray(new PhotocpyMchine[0]);
     }
     @Override
-    public Monitors getDevice(String regNum) {
+    public PhotocpyMchine getDevice(String regNum) {
         //pass query to the connection class
-        String sql = "SELECT * FROM monitors Where MonitoRegNum=?";
+        String sql = "SELECT * FROM PhotoCopyMachine Where PhotoCopyMachineRegNum=?";
 
         try {
             PreparedStatement ps = conn.getConnection().prepareStatement(sql);
@@ -113,12 +100,12 @@ public class Monitors extends Devices{
             ResultSet rs = ps.executeQuery();
 
             while (rs.next()) {
-                Monitors monitors = new Monitors();
-                monitors.setRegNum(rs.getString("regNum"));
-                monitors.setModel(rs.getString("model"));
-                monitors.setStatus(rs.getString("status"));
+                PhotocpyMchine PhotoCopyMachine = new PhotocpyMchine();
+                PhotoCopyMachine.setRegNum(rs.getString("PhotoCopyMachineRegNum"));
+                PhotoCopyMachine.setModel(rs.getString("model"));
+                PhotoCopyMachine.setStatus(rs.getString("status"));
 
-                return monitors;
+                return PhotoCopyMachine;
             }
         } catch (SQLException e) {
             throw new RuntimeException(e);
@@ -127,11 +114,10 @@ public class Monitors extends Devices{
         //return null if there is no result
         return null;
     }
-
     public boolean updateDevice(ArrayList<String> list){
         Connection connection= conn.getConnection();
         //pass query to the connection class
-        String sql="UPDATE monitors SET model=?,status=? WHERE MonitoRegNum=?";
+        String sql="UPDATE PhotoCopyMachine SET model= ?, status= ? WHERE PhotoCopyMachineRegNum=?";
         try {
             connection.setAutoCommit(false);
 
@@ -147,7 +133,7 @@ public class Monitors extends Devices{
             //Check confirmation to change
             Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
             alert.setTitle("Confirmation");
-            alert.setContentText("Update "+ i+" rows desktop registration number " +list.get(3));
+            alert.setContentText("Update "+ i+" rows Photocopy Machine registration number " +list.get(3));
 
             Optional<ButtonType> alertResult = alert.showAndWait();//wait until button press in alert box
 
@@ -173,27 +159,26 @@ public class Monitors extends Devices{
         }
         return false;
     }
-    public boolean insertDevice(ArrayList<String> list){
-        Connection connection= conn.getConnection();
+    public boolean insertDevice(ArrayList<String> list) {
+        Connection connection = conn.getConnection();
         //pass query to the connection class
-        String sql="INSERT INTO monitors (MonitoRegNum,model,status)" +
-                "VALUES (?,?,?)";
+        String sql = "INSERT INTO PhotoCopyMachine (PhotoCopyMachineRegNum,model,status) VALUES (?,?,?)";
         try {
             connection.setAutoCommit(false);
 
-            int i=1;
+            int i = 1;
             PreparedStatement ps = connection.prepareStatement(sql);
-            for(String l:list){
-                ps.setString(i,l);
+            for (String l : list) {
+                ps.setString(i, l);
                 i++;
             }
 
-            i=ps.executeUpdate();
+            i = ps.executeUpdate();
 
             //Check confirmation to change
             Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
             alert.setTitle("Confirmation");
-            alert.setContentText("Update "+ i+" rows desktop registration number " +list.getFirst());
+            alert.setContentText("Update " + i + " rows desktop registration number " + list.getFirst());
 
             Optional<ButtonType> alertResult = alert.showAndWait();//wait until button press in alert box
 
@@ -211,7 +196,7 @@ public class Monitors extends Devices{
 
         } catch (SQLException e) {
             // Rollback the transaction on error
-            Alert alert=new Alert(Alert.AlertType.ERROR);
+            Alert alert = new Alert(Alert.AlertType.ERROR);
             alert.setTitle("Error Updating Device");
             alert.setHeaderText("An error occurred while updating the device.");
             alert.setContentText(e.getMessage());
