+++ src/main/java/org/example/hakmana/view/dialogBoxes/ContactUsDialog.java
@@ -1,6 +1,15 @@
 package org.example.hakmana.view.dialogBoxes;
 
 public class ContactUsDialog {
-    public ContactUsDialog() {
+    private static ContactUsDialog instance=null;
+    private ContactUsDialog() {
+    }
+
+    public static ContactUsDialog getInstance() {
+        if(instance==null){
+            instance=new ContactUsDialog();
+            return instance;
+        }
+        return instance;
     }
 }
