+++ src/main/java/org/example/hakmana/view/scene/DeviceMngmntController.java
@@ -42,8 +42,8 @@ public class DeviceMngmntController implements Initializable {
         addComponent("Projectors",new Image(Objects.requireNonNull(getClass().getResourceAsStream("Images/projector.png"))));
         addComponent("Laptops",new Image(Objects.requireNonNull(getClass().getResourceAsStream("Images/laptopcat.png"))));
         addComponent("UPS",new Image(Objects.requireNonNull(getClass().getResourceAsStream("Images/UPS.png"))));
+        addComponent("Printers",new Image(Objects.requireNonNull(getClass().getResourceAsStream("Images/other.png"))));
         addComponent("Other Devices",new Image(Objects.requireNonNull(getClass().getResourceAsStream("Images/other.png"))));
-
         //After adding componennt reset the row and column number
         rowCount=1;
         colCount=0;
