+++ src/main/java/org/example/hakmana/view/scene/UserMngmntController.java
@@ -12,6 +12,9 @@ import org.example.hakmana.model.DatabaseConnection;
 import org.example.hakmana.model.SystemUser;
 
 import org.example.hakmana.model.SystemUser;
+import org.example.hakmana.view.dialogBoxes.CreateAccountController;
+import org.example.hakmana.view.dialogBoxes.EditAccountController;
+import org.example.hakmana.view.dialogBoxes.ShowUsersController;
 
 
 import java.io.IOException;
@@ -68,7 +71,8 @@ public class UserMngmntController implements Initializable {
     @FXML
     public void handleCreateAccountButtonAction(ActionEvent event) throws IOException {
         FXMLLoader createAccfxmlLoad = new FXMLLoader();
-        createAccfxmlLoad.setLocation(UserMngmntController.class.getResource("DialogBox/CreateAccount.fxml"));
+        createAccfxmlLoad.setLocation(org.example.hakmana.view.dialogBoxes.CreateAccountController.class.getResource("CreateAccount.fxml"));
+        createAccfxmlLoad.setController(CreateAccountController.getInstance());
         DialogPane createAccDialogPane = createAccfxmlLoad.load();
 
         Dialog<ButtonType> dialog = new Dialog<>();
@@ -81,7 +85,8 @@ public class UserMngmntController implements Initializable {
     @FXML
     protected void handleEditProfileButtonAction(ActionEvent event) throws IOException {
         FXMLLoader editAccfxmlLoad = new FXMLLoader();
-        editAccfxmlLoad.setLocation(UserMngmntController.class.getResource("DialogBox/EditProfile.fxml"));
+        editAccfxmlLoad.setLocation(org.example.hakmana.view.dialogBoxes.EditAccountController.class.getResource("EditProfile.fxml"));
+        editAccfxmlLoad.setController(EditAccountController.getInstance());
         DialogPane editAccDialogPane = editAccfxmlLoad.load();
 
         Dialog<ButtonType> dialog = new Dialog<>();
@@ -94,7 +99,8 @@ public class UserMngmntController implements Initializable {
     @FXML
     protected void handleShowUsersButtonAction(ActionEvent event) throws IOException {
         FXMLLoader showAccfxmlLoad = new FXMLLoader();
-        showAccfxmlLoad.setLocation(UserMngmntController.class.getResource("DialogBox/ShowUsers.fxml"));
+        showAccfxmlLoad.setLocation(org.example.hakmana.view.dialogBoxes.ShowUsersController.class.getResource("ShowUsers.fxml"));
+        showAccfxmlLoad.setController(ShowUsersController.getInsance());
         DialogPane showAccDialogPane = showAccfxmlLoad.load();
 
         Dialog<ButtonType> dialog = new Dialog<>();
