+++ src/main/resources/org/example/hakmana/view/dialogBoxes/AddnoteDialog.fxml
@@ -9,7 +9,7 @@
 <?import javafx.scene.layout.Pane?>
 <?import javafx.scene.text.Text?>
 
-<DialogPane fx:id="dialogpane1" expanded="true" prefHeight="446.0" prefWidth="560.0" stylesheets="@../scene/CSS/dialogPane.css" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.hakmana.view.dialogBoxes.AddNoteDialogPane">
+<DialogPane fx:id="dialogpane1" expanded="true" prefHeight="446.0" prefWidth="560.0" stylesheets="@../scene/CSS/dialogPane.css" xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1">
    <graphic>
       <Pane prefHeight="370.0" prefWidth="560.0" style="-fx-background-color: linear-gradient( #CFCDCD -9.83%, #F2F1F1 114.02%);;">
          <children>
