+++ src/main/resources/org/example/hakmana/view/dialogBoxes/EditProfile.fxml
@@ -9,7 +9,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<DialogPane xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.hakmana.view.dialogBoxes.EditAccountController">
+<DialogPane xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" >
    <content>
       <AnchorPane fx:id="parentAnchor" style="-fx-background-color: #00000021;">
           <children>
