+++ src/main/resources/org/example/hakmana/view/dialogBoxes/ShowUsers.fxml
@@ -10,7 +10,7 @@
 <?import javafx.scene.layout.VBox?>
 <?import javafx.scene.text.Font?>
 
-<DialogPane xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1" fx:controller="org.example.hakmana.view.dialogBoxes.ShowUsersController">
+<DialogPane xmlns="http://javafx.com/javafx/21" xmlns:fx="http://javafx.com/fxml/1">
    <content>
       <AnchorPane fx:id="parentAnchor" prefWidth="600.0" style="-fx-background-color: #00000021;" stylesheets="@../scene/CSS/ShowUsers.css">
           <children>
